requirements.txt
# RAG Application Dependencies
openai>=1.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
python-dotenv>=1.0.0
numpy>=1.24.0
streamlit>=1.28.0
pandas>=2.0.0
tiktoken>=0.5.0
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn>=0.24.0
pydantic>=2.6.0
